Move backend setup functions
